Merge sets of paired interactions

Sets of paired range objects (i.e., `GInteractions`) are first clustered by genomic distance with `dbscan`, then a representative interaction is selected for each cluster.

Usage

mergePairs(
  x,
  radius,
  method = "manhattan",
  column = NULL,
  selectMax = TRUE,
  pos = "center"
)

# S4 method for list_OR_SimpleList_OR_GInteractions,numeric
mergePairs(
  x,
  radius,
  method = "manhattan",
  column = NULL,
  selectMax = TRUE,
  pos = "center"
)

Arguments

x

List of `GInteractions` or `data.frame`-like objects.

radius

Numeric describing the distance in base pairs used to define a cluster or pairs.

method

Character describing the distance measure to be used. This must be one of "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski". Any unambiguous substring can be given. Default is "manhattan".

column

Character denoting the column to be used to select among clustered interactions.

selectMax

Logical. TRUE (default) uses `which.max()` to select the interaction pair. FALSE uses `which.min()`. Only applicable when `column` is specified.

pos

Positions used for clustering pairs. Must be one of "start", "end" or "center". Default is "center".

Value

Returns a `MergedGInteractions` object.

Details

Interactions are clustered into groups using the provided base pair `radius`, and distance `method` with `dbscan()`. Representative interactions are selected for each group by one of two methods. If `column` and `selectMax` arguments are provided, the representative interaction with the maximum (or minimum) value in `column` is returned for each cluster. If these parameters are missing, new ranges for each pair are returned by calculating the median of modes for each cluster.

Examples

## Load required packages
library(data.table, include.only="fread")

## Reference BEDPE files (loops called with SIP)
bedpeFiles <-
    system.file("extdata", package = "mariner") |>
    list.files(pattern = "Loops.txt", full.names = TRUE)

giList <-
    lapply(bedpeFiles, fread) |>
    lapply(as_ginteractions)

x <- mergePairs(x = giList,
                radius = 10e03,
                column = "APScoreAvg")
x
#> MergedGInteractions object with 16716 interactions and 9 metadata columns:
#>           seqnames1             ranges1     seqnames2             ranges2 |
#>               <Rle>           <IRanges>         <Rle>           <IRanges> |
#>       [1]      chr9 118645000-118650000 ---      chr9 119330000-119335000 |
#>       [2]      chr9   15280000-15285000 ---      chr9   15405000-15410000 |
#>       [3]      chr9 110180000-110185000 ---      chr9 111520000-111525000 |
#>       [4]      chr9   80375000-80380000 ---      chr9   80650000-80655000 |
#>       [5]      chr9 108380000-108385000 ---      chr9 108475000-108480000 |
#>       ...       ...                 ... ...       ...                 ... .
#>   [16712]     chr17   71425000-71430000 ---     chr17   72170000-72175000 |
#>   [16713]     chr17   28450000-28455000 ---     chr17   28660000-28665000 |
#>   [16714]     chr17   65375000-65380000 ---     chr17   65685000-65690000 |
#>   [16715]     chr17   77710000-77715000 ---     chr17   77965000-77970000 |
#>   [16716]     chr17   46880000-46885000 ---     chr17   46955000-46960000 |
#>                 color APScoreAvg ProbabilityofEnrichment RegAPScoreAvg
#>           <character>  <numeric>               <numeric>     <numeric>
#>       [1]       0,0,0    2.61103                0.986044       1.41438
#>       [2]       0,0,0    2.45301                0.982802       1.54370
#>       [3]       0,0,0    3.40635                0.996545       1.80221
#>       [4]       0,0,0    2.09352                0.946897       1.40194
#>       [5]       0,0,0    2.14182                0.936547       1.41005
#>       ...         ...        ...                     ...           ...
#>   [16712]       0,0,0    4.05887                0.997252       2.12187
#>   [16713]       0,0,0    4.47565                0.995379       2.49118
#>   [16714]       0,0,0    3.91395                0.996974       2.83091
#>   [16715]       0,0,0    3.39028                0.979869       2.42961
#>   [16716]       0,0,0    3.89043                0.982768       2.73390
#>           Avg_diffMaxNeihgboor_1 Avg_diffMaxNeihgboor_2       avg       std
#>                        <numeric>              <numeric> <numeric> <numeric>
#>       [1]               1.875585                2.12164   2.60512  0.794220
#>       [2]               1.491568                1.60777   2.73756  0.856876
#>       [3]               3.768934                3.89112   4.49898  1.492786
#>       [4]               0.818514                1.10611   2.20826  0.589976
#>       [5]               0.907892                1.01824   1.95072  0.543293
#>       ...                    ...                    ...       ...       ...
#>   [16712]               2.484298                3.15533   3.68911  1.144096
#>   [16713]               2.181075                2.63508   3.43889  0.827104
#>   [16714]               1.091847                1.96232   4.83071  0.753955
#>   [16715]               1.139229                1.15973   2.89326  0.589743
#>   [16716]               0.906871                1.43305   3.25531  0.368257
#>               value
#>           <numeric>
#>       [1]   4.27231
#>       [2]   4.06339
#>       [3]   7.84915
#>       [4]   2.93583
#>       [5]   2.75773
#>       ...       ...
#>   [16712]   5.89737
#>   [16713]   5.37762
#>   [16714]   5.80124
#>   [16715]   3.90591
#>   [16716]   4.06142
#>   -------
#>   regions: 27152 ranges and 0 metadata columns
#>   seqinfo: 23 sequences from an unspecified genome; no seqlengths