CountMatrix Class

A class for displaying dimnames associated with the count matrices resulting from pullHicMatrices() |> counts(showDimnames=TRUE).

Usage

# S4 method for CountMatrix
show(object)

Arguments

object

A CountMatrix object.

Value

A CountMatrix object (clone of DelayedArray)

A DelayedArray object with dimnames for the first two dimensions.

Details

This class is used only for attaching a "show" method.

Slots

object

InteractionArray object

Examples

## Load marinerData
if (!require("marinerData", quietly = TRUE))
    BiocManager::install("marinerData")

## Read .hic file paths
hicFiles <- c(
    marinerData::LEUK_HEK_PJA27_inter_30.hic(),
    marinerData::LEUK_HEK_PJA30_inter_30.hic()
)
#> see ?marinerData and browseVignettes('marinerData') for documentation
#> loading from cache
#> see ?marinerData and browseVignettes('marinerData') for documentation
#> loading from cache
#> Error: failed to load resource
#>   name: EH8089
#>   title: LEUK_HEK_PJA30_inter_30.hic
#>   reason: error in evaluating the argument 'conn' in selecting a method for function 'dbDisconnect': object 'info' not found
names(hicFiles) <- c("FS", "WT")
#> Error in names(hicFiles) <- c("FS", "WT"): object 'hicFiles' not found

## Read in loop pixels as GInteractions object
pixels <-
    WT_5kbLoops.txt() |>
    setNames("WT") |>
    read.table(header=TRUE) |>
    as_ginteractions(keep.extra.columns=FALSE) |>
    binPairs(binSize=100e3)
#> see ?marinerData and browseVignettes('marinerData') for documentation
#> loading from cache

## Removes the "chr" prefix for compatibility
## with the preprocessed hic files
GenomeInfoDb::seqlevelsStyle(pixels) <- 'ENSEMBL'

## Expand pixels to regions for pulling
## Hi-C submatrices
regions <- pixelsToMatrices(x=pixels, buffer=5)

## Extract 11x11 count matrices from the
## first 100 regions and 2 Hi-C files
iarr <- pullHicMatrices(x=regions[1:100],
                        files=hicFiles,
                        binSize=100e3)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'files' in selecting a method for function 'pullHicMatrices': object 'hicFiles' not found

## Display the start bin of each
## interaction in the count
## matrices
counts(iarr, showDimnames=TRUE)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'counts': object 'iarr' not found